Refactor EditEndpoint to ReconciliationService
- splits out some of the massive run logic to methods
- injects services

Bug: T282243
