+++ includes/MediaWiki/Api/EditEndpoint.php
@@ -2,21 +2,12 @@
 
 namespace MediaWiki\Extension\WikibaseReconcileEdit\MediaWiki\Api;
 
-use DataValues\StringValue;
-use MediaWiki\Extension\WikibaseReconcileEdit\EditStrategy\SimplePutStrategy;
-use MediaWiki\Extension\WikibaseReconcileEdit\InputToEntity\FullWikibaseItemInput;
-use MediaWiki\Extension\WikibaseReconcileEdit\InputToEntity\MinimalItemInput;
-use MediaWiki\Extension\WikibaseReconcileEdit\MediaWiki\ExternalLinks;
 use MediaWiki\Extension\WikibaseReconcileEdit\MediaWiki\Request\EditRequest;
 use MediaWiki\Extension\WikibaseReconcileEdit\MediaWiki\Request\MockEditDiskRequest;
 use MediaWiki\Extension\WikibaseReconcileEdit\MediaWiki\Request\UrlInputEditRequest;
+use MediaWiki\Extension\WikibaseReconcileEdit\Wikibase\ReconciliationService;
 use MediaWiki\Rest\LocalizedHttpException;
 use MediaWiki\Rest\SimpleHandler;
-use Title;
-use Wikibase\DataModel\Entity\Item;
-use Wikibase\DataModel\Entity\ItemId;
-use Wikibase\DataModel\Entity\PropertyId;
-use Wikibase\DataModel\Snak\PropertyValueSnak;
 use Wikibase\Lib\Store\EntityIdLookup;
 use Wikibase\Lib\Store\EntityRevision;
 use Wikibase\Repo\WikibaseRepo;
@@ -25,20 +16,17 @@ use Wikimedia\ParamValidator\ParamValidator;
 
 class EditEndpoint extends SimpleHandler {
 
-	public const VERSION_KEY = "wikibasereconcileedit-version";
-
 	/**
 	 * @var EntityIdLookup
 	 */
 	private $entityIdLookup;
 
-	/** @var ExternalLinks */
-	private $externalLinks;
+	private $reconciliationService;
 
 	public function __construct(
-		ExternalLinks $externalLinks
+		ReconciliationService $reconciliationService
 	) {
-		$this->externalLinks = $externalLinks;
+		$this->reconciliationService = $reconciliationService;
 	}
 
 	/**
@@ -55,8 +43,6 @@ class EditEndpoint extends SimpleHandler {
 	public function run() {
 		// TODO inject these services
 		$repo = WikibaseRepo::getDefaultInstance();
-		$propertyDataTypeLookup = $repo->getPropertyDataTypeLookup();
-		$deserializer = $repo->getBaseDataModelDeserializerFactory()->newEntityDeserializer();
 		$this->entityIdLookup = $repo->getEntityIdLookup();
 
 		// Get the request
@@ -65,129 +51,14 @@ class EditEndpoint extends SimpleHandler {
 		// Validate and process reconciliation input
 		// TODO use different services per version
 		// TODO output an object that controls the reconciliations spec?
-		$inputReconcile = $request->reconcile();
-		if ( $inputReconcile === null ) {
+		if ( $request->reconcile() === null ) {
 			throw new LocalizedHttpException(
 				MessageValue::new( 'wikibasereconcileedit-editendpoint-invalid-reconcile-json' ),
 				400
 			);
 		}
-		$supportedReconciliationVersions = [ '0.0.1' ];
-		if (
-			!array_key_exists( self::VERSION_KEY, $inputReconcile ) ||
-			!in_array( $inputReconcile[self::VERSION_KEY], $supportedReconciliationVersions )
-		) {
-			throw new LocalizedHttpException(
-				MessageValue::new( 'wikibasereconcileedit-editendpoint-unsupported-reconcile-version' )
-					->textListParams( $supportedReconciliationVersions )
-					->numParams( count( $supportedReconciliationVersions ) ),
-				400
-			);
-		}
-		if (
-			!array_key_exists( 'urlReconcile', $inputReconcile ) ||
-			!preg_match( PropertyId::PATTERN, $inputReconcile['urlReconcile'] )
-		) {
-			throw new LocalizedHttpException(
-				MessageValue::new( 'wikibasereconcileedit-editendpoint-invalid-reconcile-propertyid' )
-					->textParams( $inputReconcile[self::VERSION_KEY], 'urlReconcile' ),
-				400
-			);
-		}
-		$reconcileUrlProperty = new PropertyId( $inputReconcile['urlReconcile'] );
-		// For now this property must be of URL type
-		if ( $propertyDataTypeLookup->getDataTypeIdForProperty( $reconcileUrlProperty ) !== 'url' ) {
-			die( 'urlReconcile property must be of type url' );
-		}
+		[ $baseRevId, $toSave ] = $this->reconciliationService->reconcile( $request );
 
-		// Get Item from input
-		if ( !array_key_exists( self::VERSION_KEY, $request->entity() ) ) {
-			die( 'entity input version must be specified in key ' . self::VERSION_KEY );
-		}
-		$inputEntityVersion = $request->entity()[self::VERSION_KEY];
-		if ( $inputEntityVersion === '0.0.1/full' ) {
-			$inputEntity = ( new FullWikibaseItemInput )->getItem( $request );
-		} elseif ( $inputEntityVersion === '0.0.1/minimal' ) {
-			$inputEntity = ( new MinimalItemInput )->getItem( $request );
-		} else {
-			die( 'unknown entity input version' );
-		}
-
-		// Validate Entity
-		// Don't support references, qualifiers
-		foreach ( $inputEntity->getStatements()->toArray() as $statement ) {
-			if ( $statement->getQualifiers()->count() !== 0 || $statement->getReferences()->count() !== 0 ) {
-				die( 'Qualifiers and References are not currently supported' );
-			}
-		}
-		// Check for our reconciliation value
-		$reconciliationStatements = $inputEntity->getStatements()->getByPropertyId( $reconcileUrlProperty );
-		if ( $reconciliationStatements->count() !== 1 ) {
-			die( 'Entity must have at least one statement for the reconciliation Property' );
-		}
-		$reconciliationStatement = $reconciliationStatements->toArray()[0];
-		if ( !$reconciliationStatement->getMainSnak() instanceof PropertyValueSnak ) {
-			die( 'Reconciliation statement must be of type value ' );
-		}
-		/** @var PropertyValueSnak $reconciliationMainSnak */
-		$reconciliationMainSnak = $reconciliationStatement->getMainSnak();
-		/** @var StringValue $reconciliationDataValue */
-		$reconciliationDataValue = $reconciliationMainSnak->getDataValue();
-		$reconciliationUrl = $reconciliationDataValue->getValue();
-
-		// Find Items that use the URL
-		$itemIdsThatReferenceTheUrl = $this->getItemIdsFromPageIds(
-			$this->externalLinks->pageIdsContainingUrl( $reconciliationUrl )
-		);
-
-		// Find Items that match the URL and Property ID
-		$itemsThatReferenceTheUrlInCorrectStatement = [];
-		foreach ( $itemIdsThatReferenceTheUrl as $itemId ) {
-			/** @var Item $item */
-			$item = $repo->getEntityLookup()->getEntity( $itemId );
-			foreach ( $item->getStatements()->getByPropertyId( $reconcileUrlProperty )->toArray() as $statement ) {
-				if ( !$statement->getMainSnak() instanceof PropertyValueSnak ) {
-					continue;
-				}
-				/** @var PropertyValueSnak $mainSnak */
-				$mainSnak = $statement->getMainSnak();
-				$urlOfStatement = $mainSnak->getDataValue()->getValue();
-				if ( $urlOfStatement === $reconciliationUrl ) {
-					$itemsThatReferenceTheUrlInCorrectStatement[] = $item;
-				}
-			}
-		}
-
-		// If we have more than one item matches, something is wrong and we can't edit
-		if ( count( $itemsThatReferenceTheUrlInCorrectStatement ) > 1 ) {
-			die( 'Matched multiple Items during reconciliation :(' );
-		}
-
-		// Get our base
-		if ( count( $itemsThatReferenceTheUrlInCorrectStatement ) === 1 ) {
-			$base = $itemsThatReferenceTheUrlInCorrectStatement[0];
-			// XXX: This bit is so annoying...
-			$baseRevId = $repo->getEntityRevisionLookup()
-				->getLatestRevisionId( $base->getId() )
-				->onConcreteRevision( function ( $revId ) {
-					return $revId;
-				} )
-				->onRedirect( function () {
-					throw new \RuntimeException();
-				} )
-				->onNonexistentEntity( function () {
-					throw new \RuntimeException();
-				} )
-				->map();
-		} else {
-			$base = new Item();
-			$baseRevId = false;
-			// XXX: this is a bit evil, but needed to work around the fact we want to mint statement guids
-			$base->setId( ItemId::newFromNumber( $repo->newIdGenerator()->getNewId( 'wikibase-item' ) ) );
-		}
-
-		// And make the edit
-		$toSave = ( new SimplePutStrategy() )->apply( $base, $inputEntity );
 		$editEntity = $repo->newEditEntityFactory()->newEditEntity(
 			// TODO use a real user
 			\User::newSystemUser( 'WikibaseReconcileEditReconciliator' ),
@@ -213,21 +84,6 @@ class EditEndpoint extends SimpleHandler {
 			$response['revisionId'] = $entityRevision->getRevisionId();
 		}
 		return json_encode( $response );
-	}
-
-	/**
-	 * @param int[] $pageIds
-	 * @return ItemId[]
-	 */
-	private function getItemIdsFromPageIds( array $pageIds ) : array {
-		$itemIds = [];
-		foreach ( $pageIds as $pageId ) {
-			$entityId = $this->entityIdLookup->getEntityIdForTitle( Title::newFromID( $pageId ) );
-			if ( $entityId && $entityId instanceof ItemId ) {
-				$itemIds[] = $entityId;
-			}
-		}
-		return $itemIds;
 	}
 
 	public function needsWriteAccess() {
